requirements.txt
google-cloud-bigquery
pandas
matplotlib
seaborn
db-dtypes
